refactor(tests): split oversized test_ai_review.py by module
test_ai_review.py was 1280 lines, over the 500-line file-size taste-lint
error threshold. Split by tested module into three new files, keeping the
residual under the limit:

- test_verdict.py: get_verdict, merge_verdicts, extract_verdict, and the
  verdict-to-presentation helpers (alert type, exit code, emoji).
- test_quality_gate.py: get_failure_category, spec_validation_failed.
- test_workflow_runs.py: get_pr_changed_files, get_workflow_runs_by_pr,
  runs_overlap, get_concurrency_group_from_run, plus the _completed helper.

Residual test_ai_review.py (label/milestone parsing, formatting, workflow
lifecycle, security truncation regression) drops from 1280 to 486 lines.
All four files now pass the file-size gate (0 errors). Tests moved verbatim;
assertions unchanged. Full set: 231 tests before and after the split.

pytest.yml line 125 ran the verdict 100% coverage gate against
test_ai_review.py alone. The verdict tests moved out, so that command now
reports 63% and fails --cov-fail-under=100. Updated it to run all three
files that exercise scripts.ai_review_common.verdict (test_ai_review.py,
test_verdict.py, test_quality_gate.py), restoring 100%.

Fixes #1963
