+++ tests/test_ai_review.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import json
 import os
-import subprocess
 from pathlib import Path
 from unittest.mock import patch
 
@@ -17,93 +16,16 @@ from scripts.ai_review_common import (
     format_collapsible_section,
     format_markdown_table_row,
     format_verdict_alert,
-    get_concurrency_group_from_run,
-    get_failure_category,
     get_labels,
     get_labels_from_ai_output,
     get_milestone,
     get_milestone_from_ai_output,
-    get_pr_changed_files,
     get_verdict,
-    get_verdict_alert_type,
-    get_verdict_emoji,
-    get_verdict_exit_code,
-    get_workflow_runs_by_pr,
     initialize_ai_review,
     invoke_with_retry,
-    merge_verdicts,
-    runs_overlap,
-    spec_validation_failed,
     write_log,
     write_log_error,
 )
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _completed(stdout: str = "", stderr: str = "", rc: int = 0):
-    return subprocess.CompletedProcess(args=[], returncode=rc, stdout=stdout, stderr=stderr)
-
-
-# ---------------------------------------------------------------------------
-# Verdict parsing
-# ---------------------------------------------------------------------------
-
-
-class TestGetVerdict:
-    def test_explicit_verdict_pass(self):
-        assert get_verdict("Analysis complete. VERDICT: PASS. Good work!") == "PASS"
-
-    def test_explicit_verdict_critical_fail(self):
-        assert get_verdict("Found issues. VERDICT: CRITICAL_FAIL") == "CRITICAL_FAIL"
-
-    def test_explicit_verdict_warn(self):
-        assert get_verdict("Minor issues found. VERDICT: WARN") == "WARN"
-
-    def test_explicit_verdict_rejected(self):
-        assert get_verdict("Cannot approve. VERDICT: REJECTED") == "REJECTED"
-
-    def test_keyword_critical_fail_severe(self):
-        assert get_verdict("This has a severe issue that needs attention") == "CRITICAL_FAIL"
-
-    def test_keyword_rejected_must_fix(self):
-        assert get_verdict("You must fix this before merging") == "REJECTED"
-
-    def test_keyword_rejected_blocking(self):
-        assert get_verdict("This is a blocking issue") == "REJECTED"
-
-    def test_keyword_pass_approved(self):
-        assert get_verdict("Changes approved, good to merge") == "PASS"
-
-    def test_keyword_pass_looks_good(self):
-        assert get_verdict("Everything looks good to me") == "PASS"
-
-    def test_keyword_pass_no_issues(self):
-        assert get_verdict("I found no issues with this code") == "PASS"
-
-    def test_keyword_warn_warning(self):
-        assert get_verdict("There is a warning about potential issues") == "WARN"
-
-    def test_keyword_warn_caution(self):
-        assert get_verdict("Proceed with caution on this change") == "WARN"
-
-    def test_empty_output(self):
-        assert get_verdict("") == "CRITICAL_FAIL"
-
-    def test_none_output(self):
-        assert get_verdict("") == "CRITICAL_FAIL"
-
-    def test_whitespace_only(self):
-        assert get_verdict("   ") == "CRITICAL_FAIL"
-
-    def test_unparseable_output(self):
-        assert get_verdict("Some random text without any verdict keywords") == "CRITICAL_FAIL"
-
-    def test_explicit_verdict_overrides_keyword(self):
-        assert get_verdict("This looks good but VERDICT: CRITICAL_FAIL") == "CRITICAL_FAIL"
-
 
 # ---------------------------------------------------------------------------
 # Label parsing
@@ -141,8 +63,6 @@ class TestGetLabels:
     def test_adjacent_labels(self):
         labels = get_labels("LABEL:bug LABEL:urgent")
         assert len(labels) == 2
-
-
 # ---------------------------------------------------------------------------
 # Milestone parsing
 # ---------------------------------------------------------------------------
@@ -163,529 +83,6 @@ class TestGetMilestone:
 
     def test_milestone_with_numbers(self):
         assert get_milestone("MILESTONE: Sprint-42") == "Sprint-42"
-
-
-# ---------------------------------------------------------------------------
-# Verdict aggregation
-# ---------------------------------------------------------------------------
-
-
-class TestMergeVerdicts:
-    def test_all_pass(self):
-        assert merge_verdicts(["PASS", "PASS", "PASS"]) == "PASS"
-
-    def test_warn_with_pass(self):
-        assert merge_verdicts(["PASS", "WARN", "PASS"]) == "WARN"
-
-    def test_critical_fail_present(self):
-        assert merge_verdicts(["PASS", "CRITICAL_FAIL", "PASS"]) == "CRITICAL_FAIL"
-
-    def test_rejected_present(self):
-        assert merge_verdicts(["PASS", "REJECTED", "WARN"]) == "CRITICAL_FAIL"
-
-    def test_critical_over_warn(self):
-        assert merge_verdicts(["WARN", "CRITICAL_FAIL", "WARN"]) == "CRITICAL_FAIL"
-
-    def test_single_pass(self):
-        assert merge_verdicts(["PASS"]) == "PASS"
-
-    def test_single_critical_fail(self):
-        assert merge_verdicts(["CRITICAL_FAIL"]) == "CRITICAL_FAIL"
-
-    def test_fail_present(self):
-        assert merge_verdicts(["PASS", "FAIL", "WARN"]) == "CRITICAL_FAIL"
-
-    def test_empty_array(self):
-        # REQ-008-05 (issue #1934): empty sequence returns UNKNOWN; the caller
-        # cannot claim PASS when no axes were evaluated. Behavior changed from
-        # PASS in PR #1934.
-        assert merge_verdicts([]) == "UNKNOWN"
-
-    def test_unknown_alone(self):
-        assert merge_verdicts(["UNKNOWN"]) == "UNKNOWN"
-
-    def test_unknown_with_pass(self):
-        # UNKNOWN downgrades a would-be PASS: caller cannot claim PASS when
-        # an axis failed to evaluate.
-        assert merge_verdicts(["PASS", "UNKNOWN"]) == "UNKNOWN"
-
-    def test_unknown_with_warn(self):
-        # UNKNOWN does not override a real WARN finding.
-        assert merge_verdicts(["WARN", "UNKNOWN"]) == "WARN"
-
-    def test_unknown_with_critical(self):
-        # UNKNOWN does not override CRITICAL_FAIL.
-        assert merge_verdicts(["CRITICAL_FAIL", "UNKNOWN"]) == "CRITICAL_FAIL"
-
-    def test_all_unknown(self):
-        assert merge_verdicts(["UNKNOWN", "UNKNOWN", "UNKNOWN"]) == "UNKNOWN"
-
-    def test_unrecognized_token_returns_unknown(self):
-        # PR #1965 cluster J: previously unrecognized tokens silently fell
-        # through to PASS, undermining the UNKNOWN safety mechanism.
-        # Garbage input must produce UNKNOWN, never PASS.
-        assert merge_verdicts(["FOOBAR"]) == "UNKNOWN"
-        assert merge_verdicts(["pass"]) == "UNKNOWN"  # lowercase
-        assert merge_verdicts(["Pass"]) == "UNKNOWN"  # mixed case
-        assert merge_verdicts(["PASS", "FOOBAR"]) == "UNKNOWN"
-
-    def test_unrecognized_does_not_override_critical(self):
-        # CRITICAL_FAIL still wins over unrecognized tokens.
-        assert merge_verdicts(["FOOBAR", "CRITICAL_FAIL"]) == "CRITICAL_FAIL"
-
-    def test_unrecognized_does_not_override_warn(self):
-        # WARN still wins over unrecognized tokens.
-        assert merge_verdicts(["FOOBAR", "WARN"]) == "WARN"
-
-    def test_compliant_treated_as_pass(self):
-        # PR #1965 coderabbit Y14: COMPLIANT is a CI-valid token from the
-        # spec-validation flow; merge as PASS-equivalent.
-        assert merge_verdicts(["COMPLIANT"]) == "PASS"
-        assert merge_verdicts(["PASS", "COMPLIANT"]) == "PASS"
-
-    def test_non_compliant_treated_as_fail(self):
-        # NON_COMPLIANT is in FAIL_VERDICTS now.
-        assert merge_verdicts(["NON_COMPLIANT"]) == "CRITICAL_FAIL"
-        assert merge_verdicts(["PASS", "NON_COMPLIANT"]) == "CRITICAL_FAIL"
-
-    def test_partial_treated_as_warn(self):
-        # PARTIAL is warn-equivalent (used by spec validation).
-        assert merge_verdicts(["PARTIAL"]) == "WARN"
-        assert merge_verdicts(["PASS", "PARTIAL"]) == "WARN"
-
-    def test_fail_alone_returns_critical_fail(self):
-        # FAIL is in FAIL_VERDICTS; must collapse to CRITICAL_FAIL.
-        assert merge_verdicts(["FAIL"]) == "CRITICAL_FAIL"
-
-    def test_needs_review_alone_returns_critical_fail(self):
-        # NEEDS_REVIEW added in Issue #470: AI ambiguity treated as blocking.
-        assert merge_verdicts(["NEEDS_REVIEW"]) == "CRITICAL_FAIL"
-
-    def test_needs_review_with_pass(self):
-        assert merge_verdicts(["PASS", "NEEDS_REVIEW", "PASS"]) == "CRITICAL_FAIL"
-
-
-# Parametrized AC verification: every literal vector enumerated in REQ-008-05
-# must match. PR #1965 critic Finding 2: spec contract had ACs without
-# 1:1 verbatim test mapping.
-_REQ_008_05_AC_VECTORS = [
-    # AC enumerations from REQ-008-05 (in spec order):
-    ([], "UNKNOWN"),
-    (["UNKNOWN"], "UNKNOWN"),
-    (["PASS"], "PASS"),
-    (["PASS", "WARN"], "WARN"),
-    (["PASS", "UNKNOWN"], "UNKNOWN"),
-    (["WARN", "UNKNOWN"], "WARN"),
-    (["PASS", "WARN", "CRITICAL_FAIL"], "CRITICAL_FAIL"),
-    (["PASS", "FAIL"], "CRITICAL_FAIL"),
-    (["PASS", "REJECTED"], "CRITICAL_FAIL"),
-    (["UNKNOWN", "WARN"], "WARN"),
-    (["CRITICAL_FAIL", "UNKNOWN"], "CRITICAL_FAIL"),
-    (["UNKNOWN", "UNKNOWN"], "UNKNOWN"),
-]
-
-
-@pytest.mark.parametrize("verdicts,expected", _REQ_008_05_AC_VECTORS)
-def test_req_008_05_literal_ac_vectors(verdicts, expected):
-    """Every merge_verdicts AC vector enumerated in REQ-008-05 verifies.
-
-    Adds 1:1 spec-text-to-test traceability per PR #1965 critic Finding 2.
-    """
-    from scripts.ai_review_common.verdict import merge_verdicts as _mv
-    assert _mv(verdicts) == expected, (
-        f"REQ-008-05 AC failed: merge_verdicts({verdicts}) "
-        f"returned {_mv(verdicts)!r}, spec says {expected!r}"
-    )
-
-
-class TestExtractVerdict:
-    def test_simple_verdict_line(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: PASS") == "PASS"
-
-    def test_final_verdict_prefix(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Final verdict: WARN due to X") == "WARN"
-
-    def test_uppercase_label(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("VERDICT: CRITICAL_FAIL") == "CRITICAL_FAIL"
-
-    def test_no_match_returns_unknown(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("no verdict marker here") == "UNKNOWN"
-
-    def test_empty_input(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("") == "UNKNOWN"
-
-    def test_whitespace_only(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("   \n\t  ") == "UNKNOWN"
-
-    def test_multiline_finds_marker(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        text = "## Findings\n\nSomething went wrong.\n\nVerdict: REJECTED\n\nMore text."
-        assert extract_verdict(text) == "REJECTED"
-
-    def test_indented_marker(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("   Verdict: PASS") == "PASS"
-
-    def test_invalid_token_returns_unknown(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        # Token not in the allowed set: pattern requires whole word boundary
-        assert extract_verdict("Verdict: MAYBE") == "UNKNOWN"
-
-    def test_last_match_wins(self):
-        # PR #1965 coderabbit Y5: spec says "the response MUST contain a
-        # final line matching..." so the LAST verdict marker is canonical.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: PASS\nVerdict: WARN") == "WARN"
-
-    def test_extract_needs_review_token(self):
-        # PR #1965 coderabbit Y7: NEEDS_REVIEW is in FAIL_VERDICTS but
-        # was missing from the regex alternation; now included.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: NEEDS_REVIEW") == "NEEDS_REVIEW"
-        assert extract_verdict("Final verdict: NEEDS_REVIEW") == "NEEDS_REVIEW"
-
-    def test_extract_bracketed_verdict(self):
-        # PR #1965 copilot AA1: CI action.yml accepts `VERDICT: [PASS]`
-        # bracketed form (Issue #575 fix). extract_verdict was strict on
-        # bare tokens which would mismatch.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: [PASS]") == "PASS"
-        assert extract_verdict("Final verdict: [CRITICAL_FAIL]") == "CRITICAL_FAIL"
-        assert extract_verdict("VERDICT: [WARN]") == "WARN"
-
-    def test_lowercase_token_returns_unknown(self):
-        # PR #1965 cluster A: global IGNORECASE caused `Verdict: pass` to match
-        # PASS. Token is now case-sensitive uppercase; lowercase verdict text
-        # is malformed and returns UNKNOWN.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: pass") == "UNKNOWN"
-        assert extract_verdict("Verdict: warn") == "UNKNOWN"
-        assert extract_verdict("Verdict: critical_fail") == "UNKNOWN"
-
-    def test_mixed_case_token_returns_unknown(self):
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: Pass") == "UNKNOWN"
-        assert extract_verdict("Verdict: WaRn") == "UNKNOWN"
-
-    def test_label_case_insensitive(self):
-        # Label retains IGNORECASE: VERDICT, Verdict, verdict all match.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("verdict: PASS") == "PASS"
-        assert extract_verdict("VERDICT: WARN") == "WARN"
-        assert extract_verdict("Verdict: CRITICAL_FAIL") == "CRITICAL_FAIL"
-        assert extract_verdict("FINAL VERDICT: PASS") == "PASS"
-        assert extract_verdict("final verdict: WARN") == "WARN"
-
-    def test_fenced_code_block_does_not_override_final(self):
-        # PR #1965 coderabbit Y5 (combined with cluster F): an example
-        # verdict inside a fenced code block at the top of output cannot
-        # override the real final verdict line at the bottom. Last-match
-        # semantics make this safe regardless of whether the early example
-        # is in a code block, prose, or anywhere else.
-        from scripts.ai_review_common.verdict import extract_verdict
-        text = "```text\nVerdict: PASS\n```\n\nReal output here.\n\nVerdict: WARN"
-        assert extract_verdict(text) == "WARN"
-
-    def test_template_alternation_rejected(self):
-        # PR #1965 copilot 7k: axis prompts contain literal template lines
-        # such as `VERDICT: [PASS|WARN|CRITICAL_FAIL]`. Without the trailing
-        # `(?![|A-Z_])` lookahead the pattern matched `PASS` and silently
-        # coerced a template echo to a real verdict. The lookahead rejects
-        # any token followed by `|` (alternation marker).
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("VERDICT: [PASS|WARN|CRITICAL_FAIL]") == "UNKNOWN"
-        assert extract_verdict("Verdict: PASS|WARN") == "UNKNOWN"
-        assert extract_verdict(
-            "Final verdict: [PASS|WARN|CRITICAL_FAIL|REJECTED]"
-        ) == "UNKNOWN"
-
-    def test_template_then_real_verdict_finds_real(self):
-        # An axis prompt may quote the template AND emit the real verdict
-        # later. The template line is rejected; the real bare token wins.
-        from scripts.ai_review_common.verdict import extract_verdict
-        text = (
-            "Format: VERDICT: [PASS|WARN|CRITICAL_FAIL]\n"
-            "\n"
-            "Findings: ...\n"
-            "\n"
-            "VERDICT: WARN"
-        )
-        assert extract_verdict(text) == "WARN"
-
-    def test_token_prefix_collision_rejected(self):
-        # The lookahead also rejects unknown uppercase tokens that share a
-        # known token prefix (e.g., `PASS_THROUGH`, `WARN_LATER`). Without
-        # `(?![|A-Z_])` the alternation would silently match the prefix and
-        # drop the rest as `].?` trailing.
-        from scripts.ai_review_common.verdict import extract_verdict
-        assert extract_verdict("Verdict: PASS_THROUGH") == "UNKNOWN"
-        assert extract_verdict("Verdict: WARN_LATER") == "UNKNOWN"
-
-
-# ---------------------------------------------------------------------------
-# Failure categorization
-# ---------------------------------------------------------------------------
-
-
-class TestGetFailureCategory:
-    def test_exit_code_124_is_infrastructure(self):
-        assert get_failure_category(exit_code=124) == "INFRASTRUCTURE"
-
-    def test_exit_code_124_overrides_message(self):
-        assert (
-            get_failure_category(exit_code=124, message="Security vulnerability detected")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_rate_limit_message(self):
-        assert get_failure_category(message="rate limit exceeded") == "INFRASTRUCTURE"
-
-    def test_ratelimit_no_space(self):
-        assert get_failure_category(message="API ratelimit hit") == "INFRASTRUCTURE"
-
-    def test_timeout_message(self):
-        assert get_failure_category(message="Request timed out") == "INFRASTRUCTURE"
-
-    def test_429_error(self):
-        assert get_failure_category(message="HTTP 429 Too Many Requests") == "INFRASTRUCTURE"
-
-    def test_network_error(self):
-        assert get_failure_category(message="network error connecting to API") == "INFRASTRUCTURE"
-
-    def test_502_bad_gateway(self):
-        assert get_failure_category(stderr="502 Bad Gateway") == "INFRASTRUCTURE"
-
-    def test_503_service_unavailable(self):
-        assert get_failure_category(stderr="503 Service Unavailable") == "INFRASTRUCTURE"
-
-    def test_connection_refused(self):
-        assert get_failure_category(stderr="connection refused") == "INFRASTRUCTURE"
-
-    def test_connection_reset(self):
-        assert get_failure_category(stderr="connection reset by peer") == "INFRASTRUCTURE"
-
-    def test_connection_timeout(self):
-        assert get_failure_category(stderr="connection timeout") == "INFRASTRUCTURE"
-
-    def test_copilot_cli_no_output(self):
-        assert (
-            get_failure_category(message="Copilot CLI failed (exit code 1) with no output")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_missing_copilot_access(self):
-        assert (
-            get_failure_category(message="missing Copilot access for the bot account")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_insufficient_scopes(self):
-        assert (
-            get_failure_category(message="insufficient scopes for this operation")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_empty_output_is_infrastructure(self):
-        assert get_failure_category(message="", stderr="") == "INFRASTRUCTURE"
-
-    def test_no_args_is_infrastructure(self):
-        assert get_failure_category() == "INFRASTRUCTURE"
-
-    def test_security_vulnerability_is_code_quality(self):
-        assert (
-            get_failure_category(message="Security vulnerability detected in dependencies")
-            == "CODE_QUALITY"
-        )
-
-    def test_code_quality_failure(self):
-        msg = "VERDICT: CRITICAL_FAIL - Code does not meet quality standards"
-        assert get_failure_category(message=msg) == "CODE_QUALITY"
-
-    def test_test_failure(self):
-        assert (
-            get_failure_category(message="Tests failed: 3 assertions did not pass")
-            == "CODE_QUALITY"
-        )
-
-    def test_missing_docs_is_code_quality(self):
-        assert (
-            get_failure_category(message="Missing documentation for public API")
-            == "CODE_QUALITY"
-        )
-
-    def test_message_checked_before_stderr(self):
-        assert (
-            get_failure_category(message="rate limit exceeded", stderr="")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_stderr_checked_when_message_no_match(self):
-        assert (
-            get_failure_category(message="Some unrelated message", stderr="503 Service Unavailable")
-            == "INFRASTRUCTURE"
-        )
-
-    def test_case_insensitive(self):
-        assert get_failure_category(message="RATE LIMIT EXCEEDED") == "INFRASTRUCTURE"
-        assert get_failure_category(message="Rate Limit") == "INFRASTRUCTURE"
-
-
-# ---------------------------------------------------------------------------
-# Spec validation
-# ---------------------------------------------------------------------------
-
-
-class TestSpecValidationFailed:
-    def test_trace_critical_fail(self):
-        assert spec_validation_failed("CRITICAL_FAIL", "PASS") is True
-
-    def test_trace_fail(self):
-        assert spec_validation_failed("FAIL", "PASS") is True
-
-    def test_trace_needs_review(self):
-        assert spec_validation_failed("NEEDS_REVIEW", "PASS") is True
-
-    def test_completeness_critical_fail(self):
-        assert spec_validation_failed("PASS", "CRITICAL_FAIL") is True
-
-    def test_completeness_fail(self):
-        assert spec_validation_failed("PASS", "FAIL") is True
-
-    def test_completeness_partial(self):
-        assert spec_validation_failed("PASS", "PARTIAL") is True
-
-    def test_completeness_needs_review(self):
-        assert spec_validation_failed("PASS", "NEEDS_REVIEW") is True
-
-    def test_both_pass(self):
-        assert spec_validation_failed("PASS", "PASS") is False
-
-    def test_trace_warn_completeness_pass(self):
-        assert spec_validation_failed("WARN", "PASS") is False
-
-    def test_trace_pass_completeness_warn(self):
-        assert spec_validation_failed("PASS", "WARN") is False
-
-    def test_both_warn(self):
-        assert spec_validation_failed("WARN", "WARN") is False
-
-    def test_both_fail(self):
-        assert spec_validation_failed("FAIL", "FAIL") is True
-
-    def test_warn_with_partial(self):
-        assert spec_validation_failed("WARN", "PARTIAL") is True
-
-    def test_empty_verdicts(self):
-        assert spec_validation_failed("", "") is False
-
-    def test_unknown_verdicts(self):
-        assert spec_validation_failed("UNKNOWN", "UNKNOWN") is False
-
-    def test_case_insensitive(self):
-        assert spec_validation_failed("fail", "pass") is True
-        assert spec_validation_failed("FAIL", "PASS") is True
-        assert spec_validation_failed("Fail", "Pass") is True
-
-    def test_trace_failure_with_completeness_pass(self):
-        assert spec_validation_failed("CRITICAL_FAIL", "PASS") is True
-
-    def test_completeness_failure_with_trace_pass(self):
-        assert spec_validation_failed("PASS", "CRITICAL_FAIL") is True
-
-
-# ---------------------------------------------------------------------------
-# Formatting: verdict alert type
-# ---------------------------------------------------------------------------
-
-
-class TestGetVerdictAlertType:
-    def test_pass(self):
-        assert get_verdict_alert_type("PASS") == "TIP"
-
-    def test_compliant(self):
-        assert get_verdict_alert_type("COMPLIANT") == "TIP"
-
-    def test_warn(self):
-        assert get_verdict_alert_type("WARN") == "WARNING"
-
-    def test_partial(self):
-        assert get_verdict_alert_type("PARTIAL") == "WARNING"
-
-    def test_critical_fail(self):
-        assert get_verdict_alert_type("CRITICAL_FAIL") == "CAUTION"
-
-    def test_rejected(self):
-        assert get_verdict_alert_type("REJECTED") == "CAUTION"
-
-    def test_fail(self):
-        assert get_verdict_alert_type("FAIL") == "CAUTION"
-
-    def test_unknown(self):
-        assert get_verdict_alert_type("SOMETHING_ELSE") == "NOTE"
-
-
-# ---------------------------------------------------------------------------
-# Formatting: verdict exit code
-# ---------------------------------------------------------------------------
-
-
-class TestGetVerdictExitCode:
-    def test_pass_returns_0(self):
-        assert get_verdict_exit_code("PASS") == 0
-
-    def test_warn_returns_0(self):
-        assert get_verdict_exit_code("WARN") == 0
-
-    def test_critical_fail_returns_1(self):
-        assert get_verdict_exit_code("CRITICAL_FAIL") == 1
-
-    def test_rejected_returns_1(self):
-        assert get_verdict_exit_code("REJECTED") == 1
-
-    def test_fail_returns_1(self):
-        assert get_verdict_exit_code("FAIL") == 1
-
-    def test_unknown_returns_0(self):
-        assert get_verdict_exit_code("UNKNOWN") == 0
-
-
-# ---------------------------------------------------------------------------
-# Formatting: verdict emoji
-# ---------------------------------------------------------------------------
-
-
-class TestGetVerdictEmoji:
-    def test_pass(self):
-        assert get_verdict_emoji("PASS") == "\u2705"
-
-    def test_compliant(self):
-        assert get_verdict_emoji("COMPLIANT") == "\u2705"
-
-    def test_warn(self):
-        assert get_verdict_emoji("WARN") == "\u26a0\ufe0f"
-
-    def test_partial(self):
-        assert get_verdict_emoji("PARTIAL") == "\u26a0\ufe0f"
-
-    def test_critical_fail(self):
-        assert get_verdict_emoji("CRITICAL_FAIL") == "\u274c"
-
-    def test_rejected(self):
-        assert get_verdict_emoji("REJECTED") == "\u274c"
-
-    def test_fail(self):
-        assert get_verdict_emoji("FAIL") == "\u274c"
-
-    def test_unknown(self):
-        assert get_verdict_emoji("UNKNOWN") == "\u2754"
-
-
 # ---------------------------------------------------------------------------
 # Formatting: collapsible section
 # ---------------------------------------------------------------------------
@@ -705,8 +102,6 @@ class TestFormatCollapsibleSection:
         assert "Line 1" in result
         assert "Line 2" in result
         assert "Line 3" in result
-
-
 # ---------------------------------------------------------------------------
 # Formatting: verdict alert
 # ---------------------------------------------------------------------------
@@ -737,8 +132,6 @@ class TestFormatVerdictAlert:
     def test_unknown_note(self):
         result = format_verdict_alert("UNKNOWN")
         assert "[!NOTE]" in result
-
-
 # ---------------------------------------------------------------------------
 # Formatting: markdown table row
 # ---------------------------------------------------------------------------
@@ -753,8 +146,6 @@ class TestFormatMarkdownTableRow:
 
     def test_many_columns(self):
         assert format_markdown_table_row(["A", "B", "C", "D", "E"]) == "| A | B | C | D | E |"
-
-
 # ---------------------------------------------------------------------------
 # Formatting: JSON escaping
 # ---------------------------------------------------------------------------
@@ -773,8 +164,6 @@ class TestConvertToJsonEscaped:
 
     def test_plain_string(self):
         assert convert_to_json_escaped("test") == '"test"'
-
-
 # ---------------------------------------------------------------------------
 # Workflow: initialization
 # ---------------------------------------------------------------------------
@@ -799,8 +188,6 @@ class TestInitializeAIReview:
             result = initialize_ai_review()
         assert result == target
         assert Path(target).exists()
-
-
 # ---------------------------------------------------------------------------
 # Workflow: retry logic
 # ---------------------------------------------------------------------------
@@ -844,8 +231,6 @@ class TestInvokeWithRetry:
                 invoke_with_retry(_always_fail, max_retries=3, initial_delay=1)
 
         assert delays == [1, 2]
-
-
 # ---------------------------------------------------------------------------
 # Workflow: logging
 # ---------------------------------------------------------------------------
@@ -858,8 +243,6 @@ class TestWriteLog:
         with caplog.at_level(logging.INFO):
             write_log("Test message")
         assert "Test message" in caplog.text
-
-
 class TestWriteLogError:
     def test_logs_error(self, caplog):
         import logging
@@ -867,8 +250,6 @@ class TestWriteLogError:
         with caplog.at_level(logging.ERROR):
             write_log_error("Failure message")
         assert "ERROR: Failure message" in caplog.text
-
-
 # ---------------------------------------------------------------------------
 # Workflow: environment validation
 # ---------------------------------------------------------------------------
@@ -896,173 +277,6 @@ class TestAssertEnvironmentVariables:
         with patch.dict(os.environ, {"EMPTY_VAR": ""}):
             with pytest.raises(RuntimeError, match="EMPTY_VAR"):
                 assert_environment_variables(["EMPTY_VAR"])
-
-
-# ---------------------------------------------------------------------------
-# Workflow: PR changed files
-# ---------------------------------------------------------------------------
-
-
-class TestGetPRChangedFiles:
-    def test_returns_filtered_files(self):
-        stdout = "src/main.py\nREADME.md\nsrc/utils.py\n"
-        with patch.dict(os.environ, {"GITHUB_REPOSITORY": "owner/repo"}):
-            with patch("subprocess.run", return_value=_completed(stdout=stdout)):
-                result = get_pr_changed_files(123, pattern=r"\.py$")
-        assert result == ["src/main.py", "src/utils.py"]
-
-    def test_returns_all_when_no_pattern(self):
-        stdout = "a.py\nb.md\n"
-        with patch.dict(os.environ, {"GITHUB_REPOSITORY": "owner/repo"}):
-            with patch("subprocess.run", return_value=_completed(stdout=stdout)):
-                result = get_pr_changed_files(42)
-        assert len(result) == 2
-
-    def test_returns_empty_on_failure(self):
-        with patch.dict(os.environ, {"GITHUB_REPOSITORY": "owner/repo"}):
-            with patch("subprocess.run", return_value=_completed(rc=1, stderr="err")):
-                result = get_pr_changed_files(1)
-        assert result == []
-
-
-# ---------------------------------------------------------------------------
-# Workflow: workflow run analysis
-# ---------------------------------------------------------------------------
-
-
-class TestGetWorkflowRunsByPR:
-    def test_returns_filtered_runs(self):
-        runs = [
-            {"name": "quality-gate", "pull_requests": [{"number": 42}]},
-            {"name": "other", "pull_requests": [{"number": 99}]},
-        ]
-        with patch(
-            "subprocess.run",
-            return_value=_completed(stdout=json.dumps(runs)),
-        ):
-            result = get_workflow_runs_by_pr(42, repository="owner/repo")
-        assert len(result) == 1
-        assert result[0]["name"] == "quality-gate"
-
-    def test_filters_by_workflow_name(self):
-        runs = [
-            {"name": "ai-quality-gate", "pull_requests": [{"number": 42}]},
-            {"name": "label-pr", "pull_requests": [{"number": 42}]},
-        ]
-        with patch(
-            "subprocess.run",
-            return_value=_completed(stdout=json.dumps(runs)),
-        ):
-            result = get_workflow_runs_by_pr(42, workflow_name="quality", repository="o/r")
-        assert len(result) == 1
-
-    def test_raises_on_api_failure(self):
-        with patch(
-            "subprocess.run",
-            return_value=_completed(rc=1, stderr="API error"),
-        ):
-            with pytest.raises(RuntimeError, match="Failed to get workflow runs"):
-                get_workflow_runs_by_pr(1, repository="o/r")
-
-    def test_invalid_json_response_raises(self):
-        with patch(
-            "subprocess.run",
-            return_value=_completed(stdout="not json"),
-        ):
-            with pytest.raises(RuntimeError, match="Invalid JSON"):
-                get_workflow_runs_by_pr(1, repository="o/r")
-
-
-class TestRunsOverlap:
-    def test_overlapping_runs(self):
-        run1 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T01:00:00Z"}
-        run2 = {"created_at": "2026-01-01T00:30:00Z", "updated_at": "2026-01-01T01:30:00Z"}
-        assert runs_overlap(run1, run2) is True
-
-    def test_non_overlapping_runs(self):
-        run1 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T01:00:00Z"}
-        run2 = {"created_at": "2026-01-01T02:00:00Z", "updated_at": "2026-01-01T03:00:00Z"}
-        assert runs_overlap(run1, run2) is False
-
-    def test_run2_starts_exactly_at_run1_end(self):
-        # Boundary touch (run1.end == run2.start) is NOT overlap.
-        # Half-open interval semantics: [start, end) -- the endpoint is exclusive.
-        run1 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T01:00:00Z"}
-        run2 = {"created_at": "2026-01-01T01:00:00Z", "updated_at": "2026-01-01T02:00:00Z"}
-        assert runs_overlap(run1, run2) is False
-
-    def test_run1_starts_inside_run2(self):
-        # Symmetric case: run1 starts inside run2. Previously a false-negative
-        # because the old implementation only checked `run2_start` between
-        # `run1_start` and `run1_end`.
-        run1 = {"created_at": "2026-01-01T00:30:00Z", "updated_at": "2026-01-01T01:30:00Z"}
-        run2 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T01:00:00Z"}
-        assert runs_overlap(run1, run2) is True
-
-    def test_run1_fully_contains_run2(self):
-        run1 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T02:00:00Z"}
-        run2 = {"created_at": "2026-01-01T00:30:00Z", "updated_at": "2026-01-01T01:30:00Z"}
-        assert runs_overlap(run1, run2) is True
-
-    def test_run2_fully_contains_run1(self):
-        run1 = {"created_at": "2026-01-01T00:30:00Z", "updated_at": "2026-01-01T01:30:00Z"}
-        run2 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T02:00:00Z"}
-        assert runs_overlap(run1, run2) is True
-
-    def test_run1_starts_exactly_at_run2_end(self):
-        # Symmetric boundary touch.
-        run1 = {"created_at": "2026-01-01T01:00:00Z", "updated_at": "2026-01-01T02:00:00Z"}
-        run2 = {"created_at": "2026-01-01T00:00:00Z", "updated_at": "2026-01-01T01:00:00Z"}
-        assert runs_overlap(run1, run2) is False
-
-
-class TestGetConcurrencyGroupFromRun:
-    def test_quality_gate_pr(self):
-        run = {
-            "name": "ai-quality-gate",
-            "event": "pull_request",
-            "pull_requests": [{"number": 42}],
-            "head_branch": "feat/test",
-        }
-        assert get_concurrency_group_from_run(run) == "ai-quality-42"
-
-    def test_spec_validation_pr(self):
-        run = {
-            "name": "spec-validation",
-            "event": "pull_request",
-            "pull_requests": [{"number": 10}],
-            "head_branch": "feat/spec",
-        }
-        assert get_concurrency_group_from_run(run) == "spec-validation-10"
-
-    def test_label_pr(self):
-        run = {
-            "name": "label-pr",
-            "event": "pull_request",
-            "pull_requests": [{"number": 5}],
-            "head_branch": "feat/label",
-        }
-        assert get_concurrency_group_from_run(run) == "label-pr-5"
-
-    def test_default_prefix_for_unknown_workflow(self):
-        run = {
-            "name": "custom-workflow",
-            "event": "pull_request",
-            "pull_requests": [{"number": 7}],
-            "head_branch": "feat/x",
-        }
-        assert get_concurrency_group_from_run(run) == "pr-validation-7"
-
-    def test_fallback_without_pr(self):
-        run = {
-            "name": "nightly-build",
-            "event": "schedule",
-            "pull_requests": [],
-            "head_branch": "main",
-        }
-        assert get_concurrency_group_from_run(run) == "nightly-build-main"
-
-
 # ---------------------------------------------------------------------------
 # Security-hardened JSON parsing: labels
 # ---------------------------------------------------------------------------
@@ -1148,8 +362,6 @@ class TestGetLabelsFromAIOutput:
         assert "bug" in labels
         assert "enhancement" in labels
         assert "evil; rm -rf /" not in labels
-
-
 # ---------------------------------------------------------------------------
 # Security-hardened JSON parsing: milestones
 # ---------------------------------------------------------------------------
@@ -1189,8 +401,6 @@ class TestGetMilestoneFromAIOutput:
     def test_rejects_over_50_chars(self):
         long = "v" + "1" * 50
         assert get_milestone_from_ai_output(f'{{"milestone":"{long}"}}') is None
-
-
 # ---------------------------------------------------------------------------
 # Integration: complete AI output parsing
 # ---------------------------------------------------------------------------
@@ -1230,8 +440,6 @@ class TestJSONParsingIntegration:
         for inp in malicious:
             get_labels_from_ai_output(inp)
             get_milestone_from_ai_output(inp)
-
-
 # ---------------------------------------------------------------------------
 # Regression: security review output truncation (#2006)
 # ---------------------------------------------------------------------------
@@ -1251,8 +459,6 @@ _TRUNCATED_SECURITY_OUTPUT = (
     "**Location**: complete_session_log.py:383-399\n"
     "The broad `except Exception`"
 )
-
-
 class TestSecurityTruncationRegression:
     """#2006: the security prompt now emits the VERDICT on the first line, so a
     review truncated by the output budget is still parseable."""
